fix: add mobile diffusion support, move executor to shared, bump test timeouts
- Move diffusion executor from desktop/ to shared/ (no platform-specific APIs)
- Add skipPreDownload to desktop diffusion resource (companion models resolve at load time)
- Add mobile consumer: SD 2.1 Q8_0 model, device gpu, threads 4, prediction v, vae_on_cpu true
- Bump test timeouts: 300s default, 600s for batch/seed tests
- Fix DownloadExecutor() constructor call (takes no args)
